+++ src/index.ts
@@ -1,7 +1,8 @@
 // Types
-export type { CommonOptions } from './util/api-request'
+export type { CommonOptions } from './util/common'
 export * from './types/base-types'
 export * from './types/rest-types'
 
-// Util
-export { createRepo } from './util/create-repo'
+// Client
+export type { SyncClient } from './client/sync-client'
+export { makeSyncClient } from './client/sync-client'