+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # edge-sync-client
 
 ## Unreleased
+
 ### Fixed
 - API change to allow `hash` field to be optional in `GetStoreResponse`
+
+### Changed
+- New `makeSyncClient` and `SyncClient` interface replaces `createRepo` method
 
 ## 0.1.0 (2021-07-9)